finish ex18

+++ src/main/java/com/xpanxion/java/springboot/da1/demo/controller/student3/MemberController3.java
@@ -16,7 +16,7 @@ public class MemberController3 {
     private GymService3 gymService3;
 
     @PostMapping("student3/api/v1/gym/{gymId}/member")
-    public Member3 addMember(@RequestBody Member3 member3, @PathVariable("gymId") Long gymId){
+    public Member3 addMember(@RequestBody Member3 member3, @PathVariable("gymId") int gymId){
         member3.setGymId(gymId);
         return gymService3.addMember(member3);
     }
