+++ src/main/java/com/xpanxion/java/springboot/da1/demo/model/student3/Member3.java
@@ -10,6 +10,7 @@ public class Member3 {
     //Data members
     //
 
+
     @Id
     @SequenceGenerator(
             name = "member_sequence",
@@ -20,8 +21,8 @@ public class Member3 {
             strategy = GenerationType.SEQUENCE,
             generator = "member_sequence"
     )
-    private long memberId;
-    private long gymId;
+    private int memberId;
+    private int gymId;
     private String firstName;
     private String lastName;
     private String dateOfBirth;
@@ -33,8 +34,8 @@ public class Member3 {
     //
 
     public Member3(){
-        this.memberId = -1L;
-        this.gymId = -1L;
+        this.memberId = -1;
+        this.gymId = -1;
         this.firstName = "";
         this.lastName = "";
         this.dateOfBirth = "";
@@ -42,9 +43,9 @@ public class Member3 {
         this.subEndDate = "";
     }
 
-    public Member3(Long memberId, String firstName, String lastName, String dateOfBirth, String subStartDate, String subEndDate){
+    public Member3(Integer memberId, String firstName, String lastName, String dateOfBirth, String subStartDate, String subEndDate){
         this.memberId = memberId;
-        this.gymId = -1L;
+        this.gymId = -1;
         this.firstName = firstName;
         this.lastName = lastName;
         this.dateOfBirth = dateOfBirth;
@@ -57,19 +58,19 @@ public class Member3 {
     //
 
 
-    public long getMemberId() {
+    public int getMemberId() {
         return memberId;
     }
 
-    public void setMemberId(long memberId) {
+    public void setMemberId(int memberId) {
         this.memberId = memberId;
     }
 
-    public long getGymId() {
+    public int getGymId() {
         return gymId;
     }
 
-    public void setGymId(long gymId) {
+    public void setGymId(int gymId) {
         this.gymId = gymId;
     }
